Remove caching at service layer
Moved into the store for now, though the implementation is totally
wrong there as well. This is currently really just used since the backend
data is random right now, to prevent the submissions changing every time
you pick a form. In real usage, we'd probably want to just reload the backend
data in those cases to prevent stale data.